add delete route

+++ server/server.js
@@ -13,10 +13,10 @@ var app = express();
 app.use(bodyParser.json());
 
 
-app.post('/todos',function(req, res){
-	console.log(req.body);
+app.post('/todos',authenticate, function(req, res){
 	var newTodo = new Todo({
 		text: req.body.text,
+		_creator: req.user._id
 	});
 	newTodo.save().then(function(doc){
 		res.send(doc);
@@ -25,8 +25,10 @@ app.post('/todos',function(req, res){
 	});
 });
 
-app.get('/todos', function(req, res){
-	Todo.find().then(function(todos){
+app.get('/todos', authenticate, function(req, res){
+	Todo.find({
+		_creator: req.user._id
+	}).then(function(todos){
 		res.send({
 			todos: todos,
 			otherInfo: '其他信息'
@@ -37,13 +39,16 @@ app.get('/todos', function(req, res){
 });
 
 
-app.get('/todos/:id', function(req, res){
+app.get('/todos/:id', authenticate, function(req, res){
 	var id = req.params.id;
 	if( !ObjectId.isValid(id) ){
 		return res.status(404).send();
 	}
 	
-	Todo.findById(id).then(function(todo){
+	Todo.findOne({
+		_id: id,
+		_creator: req.user._id
+	}).then(function(todo){
 		if(!todo){
 			return res.status(400).send();
 		}
@@ -57,16 +62,16 @@ app.get('/todos/:id', function(req, res){
 });
 
 
-app.delete('/todos/:id', function(req, res){
+app.delete('/todos/:id',authenticate, function(req, res){
 	var id = req.params.id;
 	if( !ObjectId.isValid(id) ){
 		return res.status(404).send();
 	}
 	
-	// Todo.findOneAndRmove({})
-	// Todo.findByIdAndRmove('')
+	// Todo.findOneAndRmove({})  可以传多个参数
+	// Todo.findByIdAndRmove('') id只能传一个id参数
 	
-	Todo.findOneAndRemove({_id: id}).then(function(todo){
+	Todo.findOneAndRemove({_id: id,_creator: req.user._id}).then(function(todo){
 		if(!todo){
 			return res.status(400).send();
 		}
@@ -79,7 +84,7 @@ app.delete('/todos/:id', function(req, res){
 	
 });
 
-app.patch('/todos/:id', function(req, res){
+app.patch('/todos/:id',authenticate, function(req, res){
 	var id = req.params.id;
 	if( !ObjectId.isValid(id) ){
 		return res.status(404).send('id 错误');
@@ -94,7 +99,7 @@ app.patch('/todos/:id', function(req, res){
 		body.completedAt = null;
 	}
 	
-	Todo.findByIdAndUpdate(id, {$set: body}, {new: true}).then(function(todo){
+	Todo.findOneAndUpdate({_id: id, _creator: req.user._id}, {$set: body}, {new: true}).then(function(todo){
 		if(!todo){
 			return res.status(404).send('没有找到');
 		}
@@ -112,10 +117,9 @@ app.patch('/todos/:id', function(req, res){
 
 
 app.post('/users',function(req, res){
-	console.log(req.body);
 	
 	var body = _.pick(req.body, ['email','password']);
-	
+	console.log(body);
 	var user = new User(body);
 	
 	
@@ -134,8 +138,28 @@ app.post('/users',function(req, res){
 
 
 app.get('/users/me',authenticate, function(req, res){
-	console.log(req);
 	res.send(req.user);
+});
+
+app.post('/users/login', function(req, res){
+	var body = _.pick(req.body, ['email', 'password'])
+	// res.send(body);
+	User.findByCredentials(body.email, body.password).then(function(user){
+		return user.generateAuthToken().then(function(token){
+			res.header('x-auth', token).send(user);
+		})
+	}).catch(function(err){
+		res.status(400).send(err);
+	})
+});
+
+
+app.delete('/users/me/token', authenticate, function(req, res){
+	req.user.removeToken(req.token).then(function(){
+		res.status(200).send();
+	},function(){
+		res.status(400).send();
+	})
 });
